cibil_analyzer/requirements.txt
flask
flask-cors
gunicorn
python-dotenv
openai
torch
numpy
scikit-learn
PyPDF2
